위치 검색 api 연동

+++ NusangJspProject/src/com/nusang/data/Location.java
@@ -7,8 +7,8 @@ import lombok.Data;
 @Data
 public class Location {
 
-	private float latitude;
-	private float longtitude;
+	private double latitude;
+	private double longtitude;
 	private String region_1;
 	private String region_2;
 	private String region_3;
